requirements.txt
Pillow
extcolors
pytest
